telegram: separate integration event
